Patents: update semicolon -> colon
A semicolon is normally used when the sentences are "independent" and not when expressly a listing (which indeed is dependent on what follows) including introductory clauses such as "Note: ". This might be a matter of style; previously I have used semicolon also in listings IF the listing contained multiple whole sentences. I make this change for aesthetic reasons after consulting ChatGPT.